updates to batching code and drqa
