+++ parlai/agents/drqa/config.py
@@ -4,11 +4,6 @@
 # LICENSE file in the root directory of this source tree. An additional grant
 # of patent rights can be found in the PATENTS file in the same directory.
 import os
-import sys
-import logging
-
-def str2bool(v):
-    return v.lower() in ('yes', 'true', 't', '1', 'y')
 
 
 def add_cmdline_args(parser):
